Rename irods_api to irods
