+++ dtp/__init__.py
@@ -4,4 +4,4 @@ from dtp.gen3.convertor import Gen3Convertor
 from dtp.gen3.exporter import Exporter
 from dtp.gen3.querier import Querier
 from dtp.gen3.submitter import Gen3Submitter
-from dtp.irods_api.irods_api import IRODSAPI
+from dtp.irods.irods import IRODSAPI
